requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==5.7.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.23.0
six==1.15.0
toml==0.10.2
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.9
zipp==3.4.0
